add createdb support and use map to manage command handlers
The command "createdb" is now supported and will create a database.  The
service instance is provided in the supplied context and should always
be asserted as a DbService instance.

In addition, rather than an ever growing if/else statement or a large
switch, hide the command handlers in a map and initialize when the
program starts.