fix bug where only one command processed per connection
The internal goroutine of the WrappedReader was only performing a single
read on the underlying buffer.  The goal is to allow as many commands as
possible during the timeout window.

NOTE: I'm not convinced this test is deterministic enough to stand up
over time, so it may need to be revisited.